[xy] Move ApiBlockHandler to block api file.
